21/03/2019 01/18

+++ src/main/java/ua/in/bibi/ecommerceonlineshopping/controller/ProductsController.java
@@ -1,12 +1,15 @@
 package ua.in.bibi.ecommerceonlineshopping.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.web.bind.annotation.*;
 import ua.in.bibi.ecommerceonlineshopping.dto.request.ProductsRequest;
+import ua.in.bibi.ecommerceonlineshopping.dto.response.DataResponse;
 import ua.in.bibi.ecommerceonlineshopping.dto.response.ProductsResponse;
 import ua.in.bibi.ecommerceonlineshopping.exception.WrongInputException;
 import ua.in.bibi.ecommerceonlineshopping.service.ProductsService;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @CrossOrigin
@@ -17,33 +20,70 @@ public class ProductsController {
     @Autowired
     private ProductsService productsService;
 
-//    create
+    //    create
 //    @PostMapping("/product")
 //    @ResponseStatus(HttpStatus.CREATED)
-@PostMapping
-public ProductsResponse create(@RequestBody ProductsRequest productsRequest) throws WrongInputException {
-    return productsService.create(productsRequest);
+    @PostMapping
+    public ProductsResponse create(@RequestBody @Valid ProductsRequest productsRequest) throws WrongInputException {
+        return productsService.create(productsRequest);
     }
 
-//    read  //in PublicController
+    //    read
 //    @GetMapping("/products")
-@GetMapping
-public List<ProductsResponse> findAll() {
-    return productsService.findAll();
+    @GetMapping("/all")
+    public List<ProductsResponse> findAll() {
+        return productsService.findAll();
     }
 
-//    update
+    @GetMapping
+    public DataResponse<ProductsResponse> getProductsSortPaginationFindValue(
+            @RequestParam(required = false) String value,
+            @RequestParam Integer page,
+            @RequestParam Integer size,
+            @RequestParam String sortFieldName,
+            @RequestParam Sort.Direction direction) {
+        if (page >= 0) {
+            if (size > 0) {
+//                System.out.println("GET ALL BRANDS pageble");
+                return productsService.findAll(value, page, size, sortFieldName, direction);
+            } else {
+                throw new IllegalArgumentException("SIZE id must not be less than ONE!");
+            }
+        } else {
+            throw new IllegalArgumentException("PAGE id must not be less than ZERO!");
+        }
+
+    }
+
+    @GetMapping("/{id}")
+    public ProductsResponse getProductsById(@PathVariable Long id) {
+        if (id > 0) {
+            return productsService.findOneById(id);
+        } else {
+            throw new IllegalArgumentException("ID must not be less than ONE!");
+        }
+    }
+
+    //    update
 //    @PutMapping("/product")
-@PutMapping
-public ProductsResponse update(@RequestParam Long id, @RequestBody ProductsRequest productsRequest) throws WrongInputException {
-    return productsService.update(id, productsRequest);
+    @PutMapping
+    public ProductsResponse update(@RequestParam Long id, @RequestBody ProductsRequest productsRequest) throws WrongInputException {
+        if (id > 0) {
+            return productsService.update(id, productsRequest);
+        } else {
+            throw new IllegalArgumentException("ID must not be less than ONE!");
+        }
     }
 
-//    delete
+    //    delete
 //    @DeleteMapping("/product")
-@DeleteMapping
+    @DeleteMapping
     public void delete(@RequestParam Long id) throws WrongInputException {
-    productsService.delete(id);
+        if (id > 0) {
+            productsService.delete(id);
+        } else {
+            throw new IllegalArgumentException("ID must not be less than ONE!");
+        }
     }
 
 //    @GetMapping("/title/{bookTitle}")
